% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\docType{data}
\name{measurements}
\alias{measurements}
\title{measurements test data}
\format{
## `measurements`
A data frame with 10 rows and 2 columns:
\describe{
  \item{id}{Arbitrary ID}
  \item{value}{random data}
  ...
}
}
\source{
Joe Ramirez, test data for global variable issue
}
\usage{
measurements
}
\description{
measurements test data
}
\keyword{datasets}
