% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements2.R
\docType{data}
\name{measurements2}
\alias{measurements2}
\title{measurements2 test data}
\format{
## `measurements2`
A data frame with 20 rows and 2 columns:
\describe{
  \item{id}{Arbitrary ID}
  \item{value}{random data}
  ...
}
}
\source{
Joe Ramirez, test data for global variable issue
}
\usage{
measurements2
}
\description{
measurements2 test data
}
\keyword{datasets}
